requirements.txt
flask
flask-login
openpyxl
werkzeug
psycopg2-binary
gunicorn